+++ R/compareGroups.R
@@ -4,7 +4,8 @@ function (formula, data, subset, na.action = NULL, y = NULL, Xext = NULL, selec 
           simplify = TRUE, ref = 1, ref.no = NA, fact.ratio = 1, ref.y = 1, p.corrected = TRUE, compute.ratio = TRUE, 
           include.miss = FALSE, oddsratio.method = "midp", chisq.test.perm = FALSE, byrow = FALSE, chisq.test.B = 2000, 
           chisq.test.seed = NULL, Date.format = "d-mon-Y", var.equal = TRUE, conf.level = 0.95, surv=FALSE,
-          riskratio = FALSE, riskratio.method = "wald", compute.prop = FALSE, lab.missing = "'Missing'") 
+          riskratio = FALSE, riskratio.method = "wald", compute.prop = FALSE, lab.missing = "'Missing'",
+          p.trend.method = "spearman") 
 {
     tibble <- FALSE
     if (missing(formula)) 
@@ -41,8 +42,6 @@ function (formula, data, subset, na.action = NULL, y = NULL, Xext = NULL, selec 
       }
     }
     
-    #print(str(data))
-
     # remove labelled class (for incompatibility with haven and Hmisc)
     for (i in 1:ncol(data)){
       if (inherits(data[,i],"labelled"))
@@ -144,6 +143,9 @@ function (formula, data, subset, na.action = NULL, y = NULL, Xext = NULL, selec 
       else
         attr(X[,i], "label") <- names(data)[pos[i]]
     }
+    
+    if (!p.trend.method%in%c("spearman","kendall","cuzick"))
+      stop("'p.trend.method' must be either 'spearman', 'kendall' or 'cuzick'")
 
     cmd <- paste(c("compareGroups.fit(X = X, y = y, include.label = include.label, Xext = Xext, selec = ", deparse(substitute(selec)), ", 
                   method = method, timemax = timemax, alpha = alpha, min.dis = min.dis, max.ylev = max.ylev, 
@@ -152,7 +154,7 @@ function (formula, data, subset, na.action = NULL, y = NULL, Xext = NULL, selec 
                   include.miss = include.miss, oddsratio.method = oddsratio.method, chisq.test.perm = chisq.test.perm, byrow = byrow, 
                   chisq.test.B = chisq.test.B, chisq.test.seed = chisq.test.seed, Date.format = Date.format, var.equal=var.equal, 
                   conf.level=conf.level,surv=surv,riskratio=riskratio,riskratio.method=riskratio.method,compute.prop=compute.prop,
-                  lab.missing=lab.missing)"), collapse="")
+                  lab.missing=lab.missing,p.trend.method=p.trend.method)"), collapse="")
 
     ans <- eval(parse(text=cmd))
     
@@ -181,7 +183,7 @@ function (formula, data, subset, na.action = NULL, y = NULL, Xext = NULL, selec 
     attr(ans, "form") <- list()
     attr(ans, "form")$formula <- formula
     attr(ans, "form")$terms <- mt
-    attr(ans,"varnames.orig")<-names(X)  #pn ###@@@@@
+    attr(ans,"varnames.orig")<-names(X)
 
     ans
 }